requirements.txt
playwright
aiogram
python-dotenv
flask